add: constant values
