+++ 13_scientific/10_cavity_.cpp
@@ -1,0 +1,95 @@
+#include <cstdlib>
+#include <cstdio>
+#include <fstream>
+#include <vector>
+
+using namespace std;
+typedef vector<vector<float>> matrix;
+
+int main() {
+  int nx = 41;
+  int ny = 41;
+  int nt = 500;
+  int nit = 50;
+  double dx = 2. / (nx - 1);
+  double dy = 2. / (ny - 1);
+  double dt = .01;
+  double rho = 1.;
+  double nu = .02;
+
+  matrix u(ny,vector<float>(nx));
+  matrix v(ny,vector<float>(nx));
+  matrix p(ny,vector<float>(nx));
+  matrix b(ny,vector<float>(nx));
+  matrix un(ny,vector<float>(nx));
+  matrix vn(ny,vector<float>(nx));
+  matrix pn(ny,vector<float>(nx));
+  for (int j=0; j<ny; j++) {
+    for (int i=0; i<nx; i++) {
+      u[j][i] = 0;
+      v[j][i] = 0;
+      p[j][i] = 0;
+      b[j][i] = 0;
+    }
+  }
+  ofstream ufile("u.dat");
+  ofstream vfile("v.dat");
+  ofstream pfile("p.dat");
+  for (int n=0; n<nt; n++) {
+    for (int j=1; j<ny-1; j++) {
+      for (int i=1; i<nx-1; i++) {
+        // Compute b[j][i]
+      }
+    }
+    for (int it=0; it<nit; it++) {
+      for (int j=0; j<ny; j++)
+        for (int i=0; i<nx; i++)
+	  pn[j][i] = p[j][i];
+      for (int j=1; j<ny-1; j++) {
+        for (int i=1; i<nx-1; i++) {
+	  // Compute p[j][i]
+	}
+      }
+      for (int j=0; j<ny; j++) {
+        // Compute p[j][0] and p[j][nx-1]
+      }
+      for (int i=0; i<nx; i++) {
+	// Compute p[0][i] and p[ny-1][i]
+      }
+    }
+    for (int j=0; j<ny; j++) {
+      for (int i=0; i<nx; i++) {
+        un[j][i] = u[j][i];
+	vn[j][i] = v[j][i];
+      }
+    }
+    for (int j=1; j<ny-1; j++) {
+      for (int i=1; i<nx-1; i++) {
+	// Compute u[j][i] and v[j][i]
+      }
+    }
+    for (int j=0; j<ny; j++) {
+      // Compute u[j][0], u[j][nx-1], v[j][0], v[j][nx-1]
+    }
+    for (int i=0; i<nx; i++) {
+      // Compute u[0][i], u[ny-1][i], v[0][i], v[ny-1][i]
+    }
+    if (n % 10 == 0) {
+      for (int j=0; j<ny; j++)
+        for (int i=0; i<nx; i++)
+          ufile << u[j][i] << " ";
+      ufile << "\n";
+      for (int j=0; j<ny; j++)
+        for (int i=0; i<nx; i++)
+          vfile << v[j][i] << " ";
+      vfile << "\n";
+      for (int j=0; j<ny; j++)
+        for (int i=0; i<nx; i++)
+          pfile << p[j][i] << " ";
+      pfile << "\n";
+    }
+  }
+  ufile.close();
+  vfile.close();
+  pfile.close();
+}
